add team authorizations

+++ app/Http/Controllers/Api/TeamController.php
@@ -17,6 +17,8 @@ class TeamController extends ApiController
      */
     public function index()
     {
+        $this->authorize('team.index');
+
         return TeamResource::collection((Team::paginate(10)));
     }
 
@@ -28,6 +30,8 @@ class TeamController extends ApiController
      */
     public function store(StoreTeamRequest $request)
     {
+        $this->authorize('team.store');
+
         $team = Team::create($request->validated());
 
         return new TeamResource($team);
@@ -41,6 +45,8 @@ class TeamController extends ApiController
      */
     public function show(Team $team)
     {
+        $this->authorize('team.show');
+
         return new TeamResource($team);
     }
 
@@ -53,6 +59,8 @@ class TeamController extends ApiController
      */
     public function update(UpdateTeamRequest $request, Team $team)
     {
+        $this->authorize('team.update');
+
         $team->update($request->validated());
 
         return new TeamResource($team);
@@ -66,6 +74,8 @@ class TeamController extends ApiController
      */
     public function destroy(Team $team)
     {
+        $this->authorize('team.destroy');
+
         $team->delete();
 
         return $this->noContent();
